move components to separate folder
