+++ client/src/app/layout.tsx
@@ -1,7 +1,7 @@
 import "./globals.css";
 
-import Navbar from "./Navbar";
-import Footer from "./Footer";
+import Navbar from "@/components/Navbar";
+import Footer from "@/components/Footer";
 
 export default function RootLayout({ children }: { children: React.ReactNode }) {
   return (
